refactor(perception): encode the browserless-agnostic-core principle + one canonical fixture (#187)
Make the architecture load-bearing in the code so no one (or me under amnesia) wires the
core to depend on a browser:
- README.md: the Surface trait, the outlier-validation finding, and the load-bearing rule —
  the headless Rust core NEVER renders and NEVER assumes a browser/display/GPU (a rack
  instance has none). It routes perception by capability NAME to whatever peer has eyes
  (WireShape::Provided, one name N adapters). A Surface is an ADAPTER on an eye-node, not a
  core dependency. SEE routes out; REASON (probe) + CV-aids stay portable. Forbidden move:
  the core spawning Chromium or assuming a local display — it won't run on a headless rack,
  which is most production installs.
- domSurface.ts header: the same adapter-not-core-dependency note where it would otherwise
  be wired wrong (nothing here may be pulled into the core's build/startup; core never calls
  findChromium()).

Compression: extract the two near-duplicate spec scene builders into ONE canonical
testScene.ts (box-a/box-b/key-light), asserted by both sceneSurface.spec and session.spec.

8/8 perception tests green, typecheck + lint clean.
